requirements.txt
# Core dependencies for RAG chatbot
azure-functions>=1.13.0
openai>=1.0.0
qdrant-client>=1.7.0
requests>=2.31.0

# Document processing
pypdf>=3.17.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Azure Storage
azure-data-tables>=12.4.0

# NLP & Search
cohere>=5.3.0
rank-bm25>=0.2.2
tiktoken>=0.5.0

# Utilities
python-dateutil>=2.8.2
pandas==2.1.4
python-dotenv>=1.0.0

